# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API Documentation",
    "version": "1.0.0",
    "description": "API documentation for my Next.js app"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://story-teller-batrinutsilvius-projects.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token from the login endpoint"
      }
    }
  },
  "paths": {
...
    "/api/categories/languages/{language_id}": {
      "get": {
        "summary": "Gets all the categories by language",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "language_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of category translations for the given language, sorted by name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "CategoryTranslation ID"
                      },
                      "name": {
                        "type": "string",
                        "example": "Adventure"
                      },
                      "category": {
                        "type": "object",
                        "properties": {
                          "photo_url": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid language ID"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Language not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/favorites/profiles/{profile_id}/languages/{language_id}": {
      "get": {
        "summary": "Gets all the favorite stories of a profile",
        "tags": [
          "Favorites"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The profile ID"
          },
          {
            "in": "path",
            "name": "language_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The language ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of favorite stories for the profile with translation for the given language",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "Favorite ID"
                      },
                      "story": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "photo_url": {
                            "type": "string",
                            "nullable": true
                          },
                          "storyTranslations": {
                            "type": "array",
                            "description": "Contains one translation for the requested language",
                            "items": {
                              "type": "object",
                              "properties": {
                                "title": {
                                  "type": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "nullable": true
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid profile ID or language ID"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Profile or language not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/profiles/{profile_id}/categories/languages/{language_id}": {
      "get": {
        "summary": "Gets all the categories of a profile",
        "tags": [
          "Profiles"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The profile ID"
          },
          {
            "in": "path",
            "name": "language_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The language ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of profile category preferences with translated names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "ProfileCategory ID"
                      },
                      "category": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "photo_url": {
                            "type": "string",
                            "nullable": true
                          },
                          "categoryTranslations": {
                            "type": "array",
                            "description": "Contains one translation for the requested language",
                            "items": {
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid profile ID or language ID"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Profile or language not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}